requirements.txt
streamlit
pandas
openai
openpyxl
matplotlib
seaborn
PyPDF2
python-pptx